Prevent caught error when using `register(accelerator, callback)` signature
Before, when calling `register(accelerator, callback)`, while shuffling the argument variables to convert from the 2-parameter to the 3-parameter signature, `win` was left holding the accelerator string, and this was passed to `title()` which expects a BrowserWindow, generating an error:
    TypeError: win.getTitle is not a function
The error was caught in `title()`, which returned "A destroyed window", leading to an inaccurate debug log.

This was not a serious issue, but even caught errors become a nuisance when debugging with breaking on caught exceptions enabled, since you have to skip over them each time you restart the program, to get to the errors you're interested in.